[SMA-230] Resolve cyclic dependencies - permissions, rights, users
