+++ solawi-bid-frontend/src/jsMain/kotlin/org/solyton/solawi/bid/module/bid/component/button/update-auction.kt
@@ -16,7 +16,7 @@ import org.evoleq.optics.storage.Storage
 import org.evoleq.optics.transform.times
 import org.solyton.solawi.bid.application.data.*
 import org.evoleq.device.data.mediaType
-import org.solyton.solawi.bid.application.permission.Right
+import org.solyton.solawi.bid.module.bid.permission.BidRight
 import org.solyton.solawi.bid.application.ui.page.auction.action.configureAuction
 import org.solyton.solawi.bid.module.bid.component.form.showUpdateAuctionModal
 import org.solyton.solawi.bid.module.bid.data.Auction
@@ -41,7 +41,7 @@ fun UpdateAuctionButton(
     // Auction can only be configured, if no rounds have been created
     val isDisabled = (storage * auction * rounds * existRounds).emit() ||
         (storage * auction * auctionAccepted).emit()||
-        (storage * userData.get).emit().isNotGranted(Right.Auction.manage)
+        (storage * userData.get).emit().isNotGranted(BidRight.Auction.manage)
 
     StdButton(
         texts * text,
